Updated docs to reflect recent changes in Dancer2 auth options.
Dancer::Plugin::Passphrase has been successfully ported to Dancer2, and
have changed the pod to reflect that this has happened, and is no longer
a maybe. Also, while Dancer::Plugin::Auth::RBAC hasn't been ported to
Dancer2, Dancer2::Plugin::Auth::YARBAC seems like a capable option, so
have listed it in the pod as an option for developers.